third_party/WebKit/LayoutTests/fast/multicol/client-rect-after-spanner.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
    description("Test that getBoundingClientRect works as expected on content after a spanner.");
</script>
<div style="-webkit-column-count:3; -webkit-column-gap:16px; position:absolute; top:100px; left:100px; line-height:50px; width:332px;">
    <br><br><br><br><br>
    <div style="-webkit-column-span:all; height:20px;"></div>
    <div id="elm1"><br></div>
    <div id="elm2"><br></div>
    <div id="elm3"><br></div>
    <div id="elm4"><br></div>
    <div id="elm5"><br></div>
</div>
<script>
    shouldBe("document.getElementById('elm1').getBoundingClientRect().left", "100");
    shouldBe("document.getElementById('elm1').getBoundingClientRect().top", "220");
    shouldBe("document.getElementById('elm2').getBoundingClientRect().left", "100");
    shouldBe("document.getElementById('elm2').getBoundingClientRect().top", "270");
    shouldBe("document.getElementById('elm3').getBoundingClientRect().left", "216");
    shouldBe("document.getElementById('elm3').getBoundingClientRect().top", "220");
    shouldBe("document.getElementById('elm4').getBoundingClientRect().left", "216");
    shouldBe("document.getElementById('elm4').getBoundingClientRect().top", "270");
    shouldBe("document.getElementById('elm5').getBoundingClientRect().left", "332");
    shouldBe("document.getElementById('elm5').getBoundingClientRect().top", "220");
</script>
