third_party/blink/web_tests/platform/generic/fast/multicol/client-rect-after-spanner-expected.txt

Test that getBoundingClientRect works as expected on content after a spanner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById('elm1').getBoundingClientRect().left is 100
PASS document.getElementById('elm1').getBoundingClientRect().top is 220
PASS document.getElementById('elm2').getBoundingClientRect().left is 100
PASS document.getElementById('elm2').getBoundingClientRect().top is 270
PASS document.getElementById('elm3').getBoundingClientRect().left is 216
PASS document.getElementById('elm3').getBoundingClientRect().top is 220
PASS document.getElementById('elm4').getBoundingClientRect().left is 216
PASS document.getElementById('elm4').getBoundingClientRect().top is 270
PASS document.getElementById('elm5').getBoundingClientRect().left is 332
PASS document.getElementById('elm5').getBoundingClientRect().top is 220
PASS successfullyParsed is true

TEST COMPLETE
